pyproject.toml
[project]
name = "twilio-chatbot-dial-out"
version = "0.1.0"
description = "Twilio dial-out example for Pipecat"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "dotenv>=0.9.9",
  "pipecat-ai[websocket,cartesia,google,silero,deepgram,runner]",
  "pipecatcloud>=0.2.7",
  "twilio",
  "uvicorn>=0.38.0",
]
